feat: add TaskStatsDialog and TaskTableView components for task statistics and management
- Implemented TaskStatsDialog for displaying task execution statistics and audit logs.
- Added TaskTableView for rendering a table of tasks with sorting and selection capabilities.
- Created taskPageConfig for managing task-related constants and utility functions.
