+++ src/pages/tasks/Tasks.tsx
@@ -1,236 +1,87 @@
-import { memo, useState, useMemo, useEffect, useCallback } from 'react';
-import { useLocation } from 'wouter';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import {
-  Boxes,
-  Play,
-  Pause,
   AlertCircle,
-  Loader2,
-  MoreVertical,
-  Calendar,
-  BarChart3,
-  Plus,
-  Search,
-  ChevronDown,
+  Boxes,
+  CheckSquare,
   ChevronLeft,
   ChevronRight,
-  Pencil,
-  Trash2,
-  ToggleLeft,
-  ToggleRight,
-  FileText,
-  RefreshCw,
-  X,
-  TrendingUp,
-  XCircle,
-  BarChart2,
-  CheckCircle2,
-  AlertTriangle,
-  Clock,
-  Activity,
-  ListOrdered,
-  User,
-  Monitor,
-  CheckSquare,
-  ListChecks,
+  Download,
   LayoutGrid,
   List,
+  Loader2,
+  Monitor,
+  Pause,
+  Play,
+  Plus,
+  RefreshCw,
   Save,
-  Download,
+  Search,
   ShieldAlert,
-  ArrowUpDown,
+  Trash2,
+  X,
 } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  CardDescription,
-  CardFooter,
-} from '@/components/ui/card';
-import { Badge } from '@/components/ui/badge';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
+import { Card, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Checkbox } from '@/components/ui/checkbox';
 import {
   Dialog,
   DialogContent,
+  DialogDescription,
+  DialogFooter,
   DialogHeader,
   DialogTitle,
-  DialogFooter,
-  DialogDescription,
 } from '@/components/ui/dialog';
 import { Input } from '@/components/ui/input';
-import { Progress } from '@/components/ui/progress';
-import { Textarea } from '@/components/ui/textarea';
+import { BatchConfirmDialog } from '@/components/tasks/BatchConfirmDialog';
 import {
-  useTasks,
-  useRunTask,
-  useCreateTask,
-  useUpdateTask,
-  useUpdateTaskStatus,
-  useDeleteTask,
-  useCancelExecution,
-  useTaskStats,
-  useTaskAuditLogs,
-  useSchedulerStatus,
-  useBatchUpdateTaskStatus,
   useBatchDeleteTask,
   useBatchRunTask,
-  useCronPreview,
-  type Task,
-  type TaskExecution,
-  type CreateTaskInput,
-  type TaskListParams,
+  useBatchUpdateTaskStatus,
+  useCancelExecution,
+  useCreateTask,
+  useDeleteTask,
+  useRunTask,
+  useSchedulerStatus,
+  useTasks,
+  useUpdateTask,
+  useUpdateTaskStatus,
   type BatchOperationResult,
+  type CreateTaskInput,
+  type Task,
+  type TaskListParams,
 } from '@/hooks/useTasks';
-import { Checkbox } from '@/components/ui/checkbox';
 import type { CheckedState } from '@radix-ui/react-checkbox';
-import { BatchConfirmDialog } from '@/components/tasks/BatchConfirmDialog';
-import { useAllCasesForSelect, type TestCase as CaseItem, type CaseType } from '@/hooks/useCases';
 import { toast } from 'sonner';
-import { cn } from '@/lib/utils';
 import {
+  STATUS_LABELS,
+  SUCCESS_RATE_THRESHOLDS,
   TASKS_CONFIG,
   TRIGGER_TYPE_LABELS,
-  STATUS_LABELS,
-  STATUS_COLORS,
-  EXECUTION_STATUS_CONFIG,
-  SUCCESS_RATE_THRESHOLDS,
 } from '@/constants/tasks';
 import { TASK_MESSAGES, TASK_PAGE } from '@/constants/messages';
-
-type TaskTriggerType = 'manual' | 'scheduled' | 'ci_triggered';
-type TaskStatusFilter = '' | 'active' | 'paused' | 'archived';
-type TaskTriggerFilter = '' | TaskTriggerType;
-
-const TASK_STATUS_FILTER_OPTIONS: readonly TaskStatusFilter[] = [
-  '',
-  'active',
-  'paused',
-  'archived',
-];
-
-const TASK_TRIGGER_FILTER_OPTIONS: readonly TaskTriggerFilter[] = [
-  '',
-  'manual',
-  'scheduled',
-  'ci_triggered',
-];
-
-const TASK_TRIGGER_TYPE_OPTIONS: ReadonlyArray<{ value: TaskTriggerType; label: string }> = [
-  { value: 'manual', label: '手动触发' },
-  { value: 'scheduled', label: '定时触发' },
-  { value: 'ci_triggered', label: 'CI 触发' },
-];
-
-const CASE_TYPE_FILTER_OPTIONS: ReadonlyArray<{ value: CaseType | ''; label: string }> = [
-  { value: '', label: '全部' },
-  { value: 'api', label: 'API' },
-  { value: 'ui', label: 'UI' },
-  { value: 'performance', label: '性能' },
-];
-
-const AUDIT_ACTION_LABELS: Record<string, { label: string; color: string }> = {
-  created: { label: '创建任务', color: 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300' },
-  updated: { label: '更新任务', color: 'bg-slate-100 text-slate-700 dark:bg-slate-800 dark:text-slate-300' },
-  deleted: { label: '删除任务', color: 'bg-red-100 text-red-700 dark:bg-red-900/30 dark:text-red-300' },
-  status_changed: { label: '状态变更', color: 'bg-yellow-100 text-yellow-700 dark:bg-yellow-900/30 dark:text-yellow-300' },
-  manually_triggered: { label: '手动触发', color: 'bg-green-100 text-green-700 dark:bg-green-900/30 dark:text-green-300' },
-  execution_cancelled: { label: '取消执行', color: 'bg-orange-100 text-orange-700 dark:bg-orange-900/30 dark:text-orange-300' },
-  compensated: { label: '漏触补偿', color: 'bg-purple-100 text-purple-700 dark:bg-purple-900/30 dark:text-purple-300' },
-  triggered: { label: '调度触发', color: 'bg-indigo-100 text-indigo-700 dark:bg-indigo-900/30 dark:text-indigo-300' },
-  retry_scheduled: { label: '重试排队', color: 'bg-amber-100 text-amber-700 dark:bg-amber-900/30 dark:text-amber-300' },
-  permanently_failed: { label: '彻底失败', color: 'bg-red-200 text-red-800 dark:bg-red-900/50 dark:text-red-200' },
-};
-
-type TaskViewMode = 'cards' | 'table';
-type TaskTagFilter = '' | 'has_cases' | 'scheduled' | 'failed' | 'healthy';
-type TaskSortKey = 'name' | 'status' | 'trigger' | 'owner' | 'latestRun' | 'successRate';
-type SortDirection = 'asc' | 'desc';
-
-const TASK_FILTER_PRESET_KEY = 'task-management.common-filters';
-
-const TASK_TAG_FILTER_OPTIONS: ReadonlyArray<{ value: TaskTagFilter; label: string }> = [
-  { value: '', label: '全部标签' },
-  { value: 'has_cases', label: '已关联用例' },
-  { value: 'scheduled', label: '定时任务' },
-  { value: 'failed', label: '失败预警' },
-  { value: 'healthy', label: '稳定任务' },
-];
-
-const TABLE_COLUMN_OPTIONS: ReadonlyArray<{ key: TaskSortKey; label: string }> = [
-  { key: 'name', label: '任务名' },
-  { key: 'status', label: '状态' },
-  { key: 'trigger', label: '触发方式' },
-  { key: 'owner', label: '负责人' },
-  { key: 'latestRun', label: '最近运行' },
-  { key: 'successRate', label: '成功率' },
-];
-
-const TASK_STATUS_SEMANTIC = {
-  success: 'bg-emerald-100 text-emerald-700 dark:bg-emerald-900/30 dark:text-emerald-300',
-  running: 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300',
-  paused: 'bg-amber-100 text-amber-700 dark:bg-amber-900/30 dark:text-amber-300',
-  failed: 'bg-red-100 text-red-700 dark:bg-red-900/30 dark:text-red-300',
-  draft: 'bg-slate-100 text-slate-700 dark:bg-slate-800 dark:text-slate-300',
-} as const;
-
-type TaskSemanticStatus = keyof typeof TASK_STATUS_SEMANTIC;
-
-const isCheckedState = (checked: CheckedState): boolean => checked === true;
-
-const getErrorMessage = (error: unknown, fallback: string = TASK_MESSAGES.GENERIC_ERROR): string =>
-  error instanceof Error ? error.message : fallback;
-
-const parseCaseCount = (task: Task): number => {
-  if (!task.case_ids) return 0;
-  try {
-    const ids = JSON.parse(task.case_ids);
-    return Array.isArray(ids) ? ids.length : 0;
-  } catch {
-    return 0;
-  }
-};
-
-const getTaskSuccessRate = (task: Task): number | null => {
-  const latest = task.recentExecutions?.[0];
-  if (!latest?.total_cases) return null;
-  return Math.round((latest.passed_cases / latest.total_cases) * 100);
-};
-
-const getTaskSemanticStatus = (task: Task): { key: TaskSemanticStatus; label: string } => {
-  if (task.status === 'archived') return { key: 'draft', label: '草稿' };
-
-  const latest = task.recentExecutions?.[0];
-  if (latest?.status === 'running' || latest?.status === 'pending') {
-    return { key: 'running', label: '运行中' };
-  }
-  if (task.status === 'paused') return { key: 'paused', label: '暂停' };
-  if (latest?.status === 'failed') return { key: 'failed', label: '失败' };
-  // 从未执行过时（latest 为空），显示"空闲"而非"成功"，避免语义误导
-  if (!latest) return { key: 'draft', label: '空闲' };
-  return { key: 'success', label: '成功' };
-};
-
-const formatTime = (value?: string): string => {
-  if (!value) return '-';
-  const d = new Date(value);
-  if (isNaN(d.getTime())) return '-';
-  return d.toLocaleString('zh-CN', {
-    month: '2-digit',
-    day: '2-digit',
-    hour: '2-digit',
-    minute: '2-digit',
-  });
-};
-
-/* ─── 主页面 ─────────────────────────────────────────── */
+import { FilterGroup } from './components/FilterGroup';
+import { SchedulerMonitorDialog } from './components/SchedulerMonitorDialog';
+import { TaskCard } from './components/TaskCard';
+import { TaskFormDialog } from './components/TaskFormDialog';
+import { TaskStatsDialog } from './components/TaskStatsDialog';
+import { TaskTableView } from './components/TaskTableView';
+import {
+  TASK_FILTER_PRESET_KEY,
+  TASK_STATUS_FILTER_OPTIONS,
+  TASK_TAG_FILTER_OPTIONS,
+  TASK_TRIGGER_FILTER_OPTIONS,
+  getTaskSemanticStatus,
+  getTaskSuccessRate,
+  getErrorMessage,
+  isCheckedState,
+  parseCaseCount,
+  type SortDirection,
+  type TaskSortKey,
+  type TaskStatusFilter,
+  type TaskTagFilter,
+  type TaskTriggerFilter,
+  type TaskViewMode,
+} from './components/taskPageConfig';
 
 export default function Tasks() {
   // ── 筛选 & 分页状态 ──────────────────────────────────
@@ -1160,1359 +1011,5 @@ export default function Tasks() {
         </DialogContent>
       </Dialog>
     </div>
-  );
-}
-
-interface FilterGroupProps {
-  label: string;
-  options: ReadonlyArray<{ value: string; label: string }>;
-  value: string;
-  onChange: (value: string) => void;
-  compact?: boolean;
-}
-
-function FilterGroup({ label, options, value, onChange, compact = false }: FilterGroupProps) {
-  return (
-    <div className={cn('flex flex-wrap items-center gap-2', compact && 'gap-1.5')}>
-      <span className="text-xs font-medium text-slate-500">{label}</span>
-      {options.map((option) => (
-        <Button
-          key={option.value || 'all'}
-          type="button"
-          size="sm"
-          variant={value === option.value ? 'default' : 'outline'}
-          className={cn('h-8', compact && 'h-7 px-2 text-xs')}
-          onClick={() => onChange(option.value)}
-        >
-          {option.label}
-        </Button>
-      ))}
-    </div>
-  );
-}
-
-interface TaskTableViewProps {
-  tasks: Task[];
-  selectedTasks: Set<number>;
-  columnVisibility: Record<TaskSortKey, boolean>;
-  sortKey: TaskSortKey;
-  sortDirection: SortDirection;
-  onSort: (key: TaskSortKey) => void;
-  onSelectTask: (taskId: number, checked: CheckedState) => void;
-  onRunTask: (taskId: number, taskName: string) => void;
-  onToggleStatus: (task: Task) => void;
-  onEditTask: (task: Task) => void;
-  onDeleteTask: (task: Task) => void;
-}
-
-function TaskTableView({
-  tasks,
-  selectedTasks,
-  columnVisibility,
-  sortKey,
-  sortDirection,
-  onSort,
-  onSelectTask,
-  onRunTask,
-  onToggleStatus,
-  onEditTask,
-  onDeleteTask,
-}: TaskTableViewProps) {
-  const visibleColumns = TABLE_COLUMN_OPTIONS.filter((column) => columnVisibility[column.key]);
-
-  return (
-    <div className="overflow-x-auto rounded-xl border border-slate-200 bg-white shadow-sm dark:border-slate-800 dark:bg-slate-900">
-      <table className="min-w-full text-sm">
-        <thead className="bg-slate-50 dark:bg-slate-800/80">
-          <tr>
-            <th className="w-10 px-3 py-3 text-left">
-              <span className="sr-only">选择</span>
-            </th>
-            {visibleColumns.map((column) => (
-              <th key={column.key} className="px-3 py-3 text-left text-xs font-semibold uppercase tracking-wide text-slate-500">
-                <button className="inline-flex items-center gap-1 hover:text-slate-700 dark:hover:text-slate-300" onClick={() => onSort(column.key)}>
-                  {column.label}
-                  {sortKey === column.key && <ArrowUpDown className={cn('h-3.5 w-3.5', sortDirection === 'desc' && 'rotate-180')} />}
-                </button>
-              </th>
-            ))}
-            <th className="px-3 py-3 text-right text-xs font-semibold uppercase tracking-wide text-slate-500">操作</th>
-          </tr>
-        </thead>
-        <tbody>
-          {tasks.map((task) => {
-            const status = getTaskSemanticStatus(task);
-            const latest = task.recentExecutions?.[0];
-            const successRate = getTaskSuccessRate(task);
-            return (
-              <tr key={task.id} className="border-t border-slate-100 hover:bg-slate-50/80 dark:border-slate-800 dark:hover:bg-slate-800/40">
-                <td className="px-3 py-3">
-                  <Checkbox checked={selectedTasks.has(task.id)} onCheckedChange={(checked) => onSelectTask(task.id, checked)} />
-                </td>
-                {columnVisibility.name && <td className="px-3 py-3 font-medium text-slate-800 dark:text-slate-100">{task.name}</td>}
-                {columnVisibility.status && (
-                  <td className="px-3 py-3">
-                    <span className={cn('inline-flex rounded-full px-2 py-0.5 text-xs font-medium', TASK_STATUS_SEMANTIC[status.key])}>{status.label}</span>
-                  </td>
-                )}
-                {columnVisibility.trigger && <td className="px-3 py-3 text-slate-600 dark:text-slate-300">{TRIGGER_TYPE_LABELS[task.trigger_type] ?? task.trigger_type}</td>}
-                {columnVisibility.owner && <td className="px-3 py-3 text-slate-600 dark:text-slate-300">{task.created_by_name ?? '-'}</td>}
-                {columnVisibility.latestRun && <td className="px-3 py-3 text-slate-600 dark:text-slate-300">{formatTime(latest?.start_time)}</td>}
-                {columnVisibility.successRate && <td className="px-3 py-3 font-medium text-slate-700 dark:text-slate-200">{successRate == null ? '-' : `${successRate}%`}</td>}
-                <td className="px-3 py-3">
-                  <div className="flex items-center justify-end gap-1">
-                    <Button variant="ghost" size="icon" className="h-8 w-8" onClick={() => onRunTask(task.id, task.name)}>
-                      <Play className="h-4 w-4" />
-                    </Button>
-                    <Button variant="ghost" size="icon" className="h-8 w-8" onClick={() => onToggleStatus(task)}>
-                      <Pause className="h-4 w-4" />
-                    </Button>
-                    <Button variant="ghost" size="icon" className="h-8 w-8" onClick={() => onEditTask(task)}>
-                      <Pencil className="h-4 w-4" />
-                    </Button>
-                    <Button variant="ghost" size="icon" className="h-8 w-8" onClick={() => onDeleteTask(task)}>
-                      <MoreVertical className="h-4 w-4" />
-                    </Button>
-                  </div>
-                </td>
-              </tr>
-            );
-          })}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-/* ─── 任务卡片 ─────────────────────────────────────────── */
-
-interface TaskCardProps {
-  task: Task;
-  onRunTask: (taskId: number, taskName: string) => void;
-  onEditTask: (task: Task) => void;
-  onToggleStatus: (task: Task) => void;
-  onDeleteTask: (task: Task) => void;
-  onStatsTask: (task: Task) => void;
-  onCancelTask: (task: Task) => void;
-  isRunning?: boolean;
-  /** [P1] 任务是否在调度器等待队列中 */
-  isQueued?: boolean;
-  /** [P1] 在队列中的位置（1-based） */
-  queuePosition?: number;
-  isBatchMode?: boolean;
-  isSelected?: boolean;
-  onSelectTask?: (taskId: number, checked: CheckedState) => void;
-}
-
-const TaskCard = memo(function TaskCard({
-  task,
-  onRunTask,
-  onEditTask,
-  onToggleStatus,
-  onDeleteTask,
-  onStatsTask,
-  onCancelTask,
-  isRunning,
-  isQueued,
-  queuePosition,
-  isBatchMode,
-  isSelected,
-  onSelectTask,
-}: TaskCardProps) {
-  const [, navigate] = useLocation();
-  const lastExecution = task.recentExecutions?.[0];
-  const successRate =
-    lastExecution?.total_cases
-      ? Math.round((lastExecution.passed_cases / lastExecution.total_cases) * 100)
-      : null;
-
-  // 是否有正在运行的执行（用于显示「取消」按钮）
-  const hasActiveExecution = task.recentExecutions?.some(
-    (e) => e.status === 'pending' || e.status === 'running'
-  ) ?? false;
-
-  // 解析关联用例数量（useMemo 避免每次渲染重复 JSON.parse）
-  const caseCount = useMemo(() => {
-    if (!task.case_ids) return 0;
-    try {
-      const ids = JSON.parse(task.case_ids);
-      return Array.isArray(ids) ? ids.length : 0;
-    } catch {
-      return 0;
-    }
-  }, [task.case_ids]);
-
-  const handleViewReport = useCallback(() => {
-    if (task.latestRunId) {
-      navigate(`/reports/${task.latestRunId}`);
-    } else {
-      navigate('/reports');
-    }
-  }, [navigate, task.latestRunId]);
-
-  const handleRun = useCallback(() => {
-    onRunTask(task.id, task.name);
-  }, [onRunTask, task.id, task.name]);
-
-  const handleEdit = useCallback(() => {
-    onEditTask(task);
-  }, [onEditTask, task]);
-
-  const handleToggle = useCallback(() => {
-    onToggleStatus(task);
-  }, [onToggleStatus, task]);
-
-  const handleDelete = useCallback(() => {
-    onDeleteTask(task);
-  }, [onDeleteTask, task]);
-
-  const handleStats = useCallback(() => {
-    onStatsTask(task);
-  }, [onStatsTask, task]);
-
-  const handleCancel = useCallback(() => {
-    onCancelTask(task);
-  }, [onCancelTask, task]);
-
-  const handleSelect = useCallback((checked: CheckedState) => {
-    onSelectTask?.(task.id, checked);
-  }, [onSelectTask, task.id]);
-
-  return (
-    <Card className={cn(
-      "group hover:shadow-xl transition-all duration-300 border-slate-200 dark:border-slate-800 overflow-hidden flex flex-col hover:-translate-y-0.5",
-      isSelected && "ring-2 ring-blue-500 border-blue-500"
-    )}>
-      <CardHeader className="pb-4">
-        <div className="flex items-start justify-between gap-2">
-          {/* 批量模式下显示复选框 */}
-          {isBatchMode && (
-            <Checkbox
-              checked={isSelected}
-              onCheckedChange={handleSelect}
-              className="mt-1 shrink-0"
-            />
-          )}
-          <div className="space-y-1 min-w-0 flex-1">
-            <CardTitle className="text-xl font-bold group-hover:text-blue-600 transition-colors truncate">
-              {task.name}
-            </CardTitle>
-            <div className="flex flex-wrap items-center gap-2 text-xs text-slate-500">
-              {/* 状态 Badge */}
-              <span
-                className={cn(
-                  'inline-flex items-center rounded-full px-2 py-0.5 font-medium text-xs',
-                  STATUS_COLORS[task.status] ?? STATUS_COLORS.active
-                )}
-              >
-                {STATUS_LABELS[task.status] ?? task.status}
-              </span>
-            </div>
-          </div>
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="ghost" size="icon" className="h-8 w-8 shrink-0">
-                <MoreVertical className="h-4 w-4" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end">
-              <DropdownMenuItem onClick={handleEdit} className="gap-2">
-                <Pencil className="h-4 w-4" />
-                {TASK_MESSAGES.BTN_EDIT_TASK}
-              </DropdownMenuItem>
-              <DropdownMenuItem onClick={handleViewReport} className="gap-2">
-                <FileText className="h-4 w-4" />
-                {TASK_MESSAGES.BTN_VIEW_REPORT}
-              </DropdownMenuItem>
-              <DropdownMenuItem onClick={handleStats} className="gap-2">
-                <TrendingUp className="h-4 w-4" />
-                查看统计
-              </DropdownMenuItem>
-              {hasActiveExecution && (
-                <DropdownMenuItem onClick={handleCancel} className="gap-2 text-orange-600">
-                  <XCircle className="h-4 w-4" />
-                  取消运行
-                </DropdownMenuItem>
-              )}
-              <DropdownMenuSeparator />
-              <DropdownMenuItem onClick={handleToggle} className="gap-2">
-                {task.status === 'active' ? (
-                  <>
-                    <ToggleLeft className="h-4 w-4" />
-                    {TASK_MESSAGES.BTN_PAUSE_TASK}
-                  </>
-                ) : (
-                  <>
-                    <ToggleRight className="h-4 w-4" />
-                    {TASK_MESSAGES.BTN_ENABLE_TASK}
-                  </>
-                )}
-              </DropdownMenuItem>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem onClick={handleDelete} className="text-red-600 gap-2">
-                <Trash2 className="h-4 w-4" />
-                {TASK_MESSAGES.BTN_DELETE_TASK}
-              </DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </div>
-      </CardHeader>
-
-      <CardContent className="flex-1 space-y-4">
-        <p className="text-sm text-slate-600 dark:text-slate-400 line-clamp-2 min-h-[2.5rem]">
-          {task.description || TASK_MESSAGES.NO_DESCRIPTION}
-        </p>
-
-        {/* 触发类型 & 成功率 */}
-        <div className="flex items-center justify-between gap-2 text-sm">
-          <div className="flex items-center gap-2 text-slate-500">
-            <Calendar className="h-4 w-4" />
-            <span>
-              {task.trigger_type === 'scheduled'
-                ? (task.cron_expression || '定时')
-                : TRIGGER_TYPE_LABELS[task.trigger_type] ?? task.trigger_type}
-            </span>
-          </div>
-          {successRate !== null && (
-            <div className="flex items-center gap-1.5">
-              <BarChart3 className="h-4 w-4 text-slate-400" />
-              <span
-                className={cn(
-                  'font-bold',
-                  successRate >= SUCCESS_RATE_THRESHOLDS.HIGH
-                    ? 'text-green-600'
-                    : successRate >= SUCCESS_RATE_THRESHOLDS.MEDIUM
-                    ? 'text-orange-600'
-                    : 'text-red-600'
-                )}
-              >
-                {successRate}%
-              </span>
-            </div>
-          )}
-        </div>
-
-        {/* 关联用例数量 */}
-        <div className="flex items-center gap-1.5 text-xs text-slate-400">
-          <ListChecks className="h-3.5 w-3.5" />
-          {caseCount > 0 ? (
-            <span>
-              已关联 <span className="font-semibold text-slate-600 dark:text-slate-300">{caseCount}</span> 个用例
-            </span>
-          ) : (
-            <span className="italic text-amber-500">暂无关联用例</span>
-          )}
-        </div>
-
-        {/* 最近运行记录小圆点 */}
-        <div className="space-y-2">
-          <div className="flex items-center justify-between text-[10px] uppercase tracking-wider font-semibold text-slate-400">
-            <span>{TASK_PAGE.RECENT_RUNS_LABEL}</span>
-            <span>{TASK_PAGE.RECENT_RUNS_COUNT(task.recentExecutions?.length ?? 0)}</span>
-          </div>
-          <div className="flex gap-1.5 h-6 items-center">
-            {task.recentExecutions && task.recentExecutions.length > 0 ? (
-              task.recentExecutions.slice(0, TASKS_CONFIG.RECENT_EXECUTIONS_DISPLAY).map((exec) => (
-                <ExecutionStatusDot key={exec.id} execution={exec} />
-              ))
-            ) : (
-              <span className="text-xs text-slate-400 italic">{TASK_MESSAGES.NO_RECENT_EXECUTIONS}</span>
-            )}
-          </div>
-        </div>
-      </CardContent>
-
-      <CardFooter className="pt-4 border-t border-slate-100 dark:border-slate-800 bg-slate-50/50 dark:bg-slate-900/50">
-        <Button
-          onClick={handleRun}
-          disabled={isRunning || isQueued || task.status === 'archived'}
-          className={cn(
-            "w-full gap-2",
-            isQueued
-              ? "bg-amber-50 hover:bg-amber-50 text-amber-700 border-amber-300 dark:bg-amber-900/20 dark:text-amber-400 dark:border-amber-700"
-              : "bg-white hover:bg-blue-50 text-blue-600 border-blue-200 hover:border-blue-300 dark:bg-slate-800 dark:text-blue-400 dark:border-slate-700"
-          )}
-          variant="outline"
-        >
-          {isRunning ? (
-            <Loader2 className="h-4 w-4 animate-spin" />
-          ) : isQueued ? (
-            <Clock className="h-4 w-4" />
-          ) : (
-            <Play className="h-4 w-4 fill-current" />
-          )}
-          {isRunning
-            ? TASK_MESSAGES.BTN_RUNNING
-            : isQueued
-            ? `排队中 ${queuePosition ? `(第 ${queuePosition} 位)` : ''}`
-            : TASK_MESSAGES.BTN_RUN_NOW}
-        </Button>
-      </CardFooter>
-    </Card>
-  );
-});
-
-TaskCard.displayName = 'TaskCard';
-
-/* ─── 执行状态小圆点 ─────────────────────────────────────── */
-
-function ExecutionStatusDot({ execution }: { execution: TaskExecution }) {
-  // 使用 EXECUTION_STATUS_CONFIG 集中管理状态样式，避免重复定义
-  const config = EXECUTION_STATUS_CONFIG[execution.status] ?? EXECUTION_STATUS_CONFIG.pending;
-
-  return (
-    <div
-      role="img"
-      className={cn(
-        'w-3 h-3 rounded-full cursor-help transition-transform hover:scale-150',
-        config.color
-      )}
-      title={`${config.label} - ${execution.start_time ?? '未知时间'}`}
-      aria-label={`${config.label} - ${execution.start_time ?? '未知时间'}`}
-    />
-  );
-}
-
-/* ─── 任务统计弹窗 ─────────────────────────────────────────── */
-
-type StatsTab = 'stats' | 'audit';
-
-function TaskStatsDialog({ task, onClose }: { task: Task; onClose: () => void }) {
-  const [activeTab, setActiveTab] = useState<StatsTab>('stats');
-  const { data: stats, isLoading: statsLoading, error: statsError } = useTaskStats(task.id, 30);
-  const { data: auditData, isLoading: auditLoading } = useTaskAuditLogs(task.id, 50, 0);
-
-  const successRateColor = (rate: number) =>
-    rate >= 80 ? 'text-green-600' : rate >= 50 ? 'text-orange-500' : 'text-red-500';
-
-  return (
-    <Dialog open onOpenChange={(v) => !v && onClose()}>
-      <DialogContent className="max-w-2xl max-h-[85vh] overflow-hidden flex flex-col p-0 animate-in fade-in-0 zoom-in-95 duration-200">
-        <DialogHeader className="px-6 pt-6 pb-3 border-b border-slate-100 dark:border-slate-800">
-          <DialogTitle className="flex items-center gap-2">
-            <BarChart2 className="h-5 w-5 text-blue-500" />
-            {task.name}
-          </DialogTitle>
-          <DialogDescription>
-            执行统计与操作审计
-          </DialogDescription>
-        </DialogHeader>
-
-        {/* 标签页切换 */}
-        <div className="flex border-b border-slate-200 dark:border-slate-700 shrink-0 px-4">
-          <button
-            className={cn(
-              'flex items-center gap-1.5 px-4 py-2.5 text-sm font-medium border-b-2 transition-colors',
-              activeTab === 'stats'
-                ? 'border-blue-500 text-blue-600 dark:text-blue-400'
-                : 'border-transparent text-slate-500 hover:text-slate-700 dark:hover:text-slate-300'
-            )}
-            onClick={() => setActiveTab('stats')}
-          >
-            <TrendingUp className="h-4 w-4" />
-            执行统计（近30天）
-          </button>
-          <button
-            className={cn(
-              'flex items-center gap-1.5 px-4 py-2.5 text-sm font-medium border-b-2 transition-colors',
-              activeTab === 'audit'
-                ? 'border-blue-500 text-blue-600 dark:text-blue-400'
-                : 'border-transparent text-slate-500 hover:text-slate-700 dark:hover:text-slate-300'
-            )}
-            onClick={() => setActiveTab('audit')}
-          >
-            <ListOrdered className="h-4 w-4" />
-            操作审计
-            {auditData && auditData.total > 0 && (
-              <span className="ml-1 text-xs bg-slate-100 dark:bg-slate-700 text-slate-600 dark:text-slate-300 px-1.5 py-0.5 rounded-full">
-                {auditData.total}
-              </span>
-            )}
-          </button>
-        </div>
-
-        <div className="flex-1 overflow-y-auto px-6 py-4">
-          {/* ─── 执行统计标签页 ─── */}
-          {activeTab === 'stats' && (
-            <div className="">
-              {statsLoading ? (
-                <div className="flex justify-center py-12">
-                  <Loader2 className="h-8 w-8 animate-spin text-blue-500" />
-                </div>
-              ) : statsError ? (
-                <div className="flex items-center gap-2 text-red-600 py-4">
-                  <AlertCircle className="h-5 w-5" />
-                  <span>加载失败：{getErrorMessage(statsError)}</span>
-                </div>
-              ) : stats ? (
-                <div className="space-y-6">
-                  {/* 摘要卡片 */}
-                  <div className="grid grid-cols-4 gap-4">
-                    <div className="rounded-xl bg-slate-50 dark:bg-slate-800 p-4 text-center">
-                      <p className="text-2xl font-bold">{stats.summary.total}</p>
-                      <p className="text-xs text-slate-500 mt-1">总执行次数</p>
-                    </div>
-                    <div className="rounded-xl bg-green-50 dark:bg-green-900/20 p-4 text-center">
-                      <p className={`text-2xl font-bold ${successRateColor(stats.summary.successRate)}`}>
-                        {stats.summary.successRate}%
-                      </p>
-                      <p className="text-xs text-slate-500 mt-1">成功率</p>
-                    </div>
-                    <div className="rounded-xl bg-red-50 dark:bg-red-900/20 p-4 text-center">
-                      <p className="text-2xl font-bold text-red-600">{stats.summary.failedCount}</p>
-                      <p className="text-xs text-slate-500 mt-1">失败次数</p>
-                    </div>
-                    <div className="rounded-xl bg-blue-50 dark:bg-blue-900/20 p-4 text-center">
-                      <p className="text-2xl font-bold text-blue-600">
-                        {stats.summary.avgDurationSec > 0 ? `${stats.summary.avgDurationSec}s` : '-'}
-                      </p>
-                      <p className="text-xs text-slate-500 mt-1">平均耗时</p>
-                    </div>
-                  </div>
-
-                  {/* 每日成功率趋势 */}
-                  {stats.trend.length > 0 && (
-                    <div className="space-y-2">
-                      <h4 className="text-sm font-semibold text-slate-700 dark:text-slate-300 flex items-center gap-1.5">
-                        <TrendingUp className="h-4 w-4 text-blue-500" />
-                        每日成功率趋势
-                      </h4>
-                      <div className="space-y-1.5 max-h-[180px] overflow-y-auto pr-1">
-                        {stats.trend.map((item) => (
-                          <div key={item.day} className="flex items-center gap-3 text-sm">
-                            <span className="w-24 shrink-0 text-slate-500 text-xs">{item.day}</span>
-                            <Progress
-                              value={Math.max(2, item.successRate)}
-                              className={cn(
-                                'h-2 flex-1',
-                                item.successRate >= 80
-                                  ? '[&>div]:bg-green-500'
-                                  : item.successRate >= 50
-                                  ? '[&>div]:bg-orange-400'
-                                  : '[&>div]:bg-red-400'
-                              )}
-                              aria-label={`${item.day} 成功率 ${item.successRate}%`}
-                            />
-                            <span className={cn('w-10 text-right font-bold text-xs shrink-0', successRateColor(item.successRate))}>
-                              {item.successRate}%
-                            </span>
-                            <span className="text-xs text-slate-400 shrink-0">{item.total}次</span>
-                          </div>
-                        ))}
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Top 10 失败原因 */}
-                  {stats.topErrors.length > 0 && (
-                    <div className="space-y-2">
-                      <h4 className="text-sm font-semibold text-slate-700 dark:text-slate-300 flex items-center gap-1.5">
-                        <AlertTriangle className="h-4 w-4 text-orange-500" />
-                        高频失败原因 TOP 10
-                      </h4>
-                      <div className="space-y-1.5 max-h-[200px] overflow-y-auto pr-1">
-                        {stats.topErrors.map((err, idx) => (
-                          <div key={idx} className="flex items-start gap-3 text-xs rounded-lg p-2 bg-red-50 dark:bg-red-900/10 border border-red-100 dark:border-red-900/20">
-                            <span className="font-bold text-red-500 shrink-0 w-4">{idx + 1}</span>
-                            <span className="flex-1 text-slate-700 dark:text-slate-300 break-all line-clamp-2">
-                              {err.errorMessage}
-                            </span>
-                            <Badge variant="secondary" className="shrink-0 text-xs bg-red-100 dark:bg-red-900/30 text-red-600">
-                              {err.count}次
-                            </Badge>
-                          </div>
-                        ))}
-                      </div>
-                    </div>
-                  )}
-
-                  {stats.trend.length === 0 && stats.topErrors.length === 0 && (
-                    <div className="text-center py-6 text-slate-400">
-                      <CheckCircle2 className="h-12 w-12 mx-auto mb-2 opacity-30" />
-                      <p>近 30 天内暂无运行记录</p>
-                    </div>
-                  )}
-                </div>
-              ) : null}
-            </div>
-          )}
-
-          {/* ─── 审计日志标签页 ─── */}
-          {activeTab === 'audit' && (
-            <div className="">
-              {auditLoading ? (
-                <div className="flex justify-center py-12">
-                  <Loader2 className="h-8 w-8 animate-spin text-blue-500" />
-                </div>
-              ) : !auditData || auditData.data.length === 0 ? (
-                <div className="text-center py-8 text-slate-400">
-                  <ListOrdered className="h-12 w-12 mx-auto mb-2 opacity-30" />
-                  <p>暂无操作记录</p>
-                </div>
-              ) : (
-                <div className="space-y-2">
-                  <p className="text-xs text-slate-500 mb-3">
-                    共 {auditData.total} 条操作记录（显示最近 50 条）
-                  </p>
-                  <div className="space-y-2 max-h-[420px] overflow-y-auto pr-1">
-                    {auditData.data.map((entry) => {
-                      const actionConfig = AUDIT_ACTION_LABELS[entry.action] ?? {
-                        label: entry.action,
-                        color: 'bg-slate-100 text-slate-600',
-                      };
-                      return (
-                        <div
-                          key={entry.id}
-                          className="flex items-start gap-3 p-3 rounded-lg border border-slate-100 dark:border-slate-800 bg-slate-50/50 dark:bg-slate-900/50 hover:bg-slate-50 dark:hover:bg-slate-800/50 transition-colors"
-                        >
-                          {/* 操作类型标签 */}
-                          <span className={cn(
-                            'shrink-0 text-xs px-2 py-1 rounded-full font-medium',
-                            actionConfig.color
-                          )}>
-                            {actionConfig.label}
-                          </span>
-
-                          {/* 元数据摘要 */}
-                          <div className="flex-1 min-w-0">
-                            <p className="text-xs text-slate-600 dark:text-slate-400 truncate">
-                              {JSON.stringify(entry.metadata) === '{}' ? '（无附加信息）' : (
-                                Object.entries(entry.metadata)
-                                  .slice(0, 3)
-                                  .map(([k, v]) => `${k}: ${JSON.stringify(v)}`)
-                                  .join(' · ')
-                              )}
-                            </p>
-                          </div>
-
-                          {/* 操作者 + 时间 */}
-                          <div className="text-right shrink-0 space-y-0.5">
-                            <p className="text-xs text-slate-500 flex items-center gap-1 justify-end">
-                              <User className="h-3 w-3" />
-                              {entry.operatorName ?? (entry.operatorId != null ? `#${entry.operatorId}` : '系统')}
-                            </p>
-                            <p className="text-xs text-slate-400 flex items-center gap-1 justify-end">
-                              <Clock className="h-3 w-3" />
-                              {entry.createdAt
-                                ? new Date(entry.createdAt).toLocaleString('zh-CN', {
-                                    month: '2-digit', day: '2-digit',
-                                    hour: '2-digit', minute: '2-digit',
-                                  })
-                                : '-'}
-                            </p>
-                          </div>
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-            </div>
-          )}
-        </div>
-
-        <DialogFooter className="shrink-0 border-t border-slate-100 dark:border-slate-800 px-6 py-3 bg-slate-50/80 dark:bg-slate-900/60">
-          <Button variant="outline" onClick={onClose}>关闭</Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-/* ─── 调度器监控弹窗 ───────────────────────────────────────── */
-
-function SchedulerMonitorDialog({ onClose }: { onClose: () => void }) {
-  const { data: status, isLoading, error, refetch } = useSchedulerStatus();
-
-  return (
-    <Dialog open onOpenChange={(v) => !v && onClose()}>
-      <DialogContent className="max-w-lg p-0 overflow-hidden animate-in fade-in-0 zoom-in-95 duration-200">
-        <DialogHeader className="px-6 pt-6 pb-3 border-b border-slate-100 dark:border-slate-800">
-          <DialogTitle className="flex items-center gap-2">
-            <Activity className="h-5 w-5 text-blue-500" />
-            调度器实时状态
-          </DialogTitle>
-          <DialogDescription>
-            每 10 秒自动刷新 · 显示运行中、排队及已计划任务
-          </DialogDescription>
-        </DialogHeader>
-
-        <div className="px-6 py-4 max-h-[68vh] overflow-y-auto">
-        {isLoading ? (
-          <div className="flex justify-center py-8">
-            <Loader2 className="h-8 w-8 animate-spin text-blue-500" />
-          </div>
-        ) : error ? (
-          <div className="flex items-center gap-2 text-red-600 py-4">
-            <AlertCircle className="h-5 w-5" />
-            <span>加载失败：{getErrorMessage(error)}</span>
-          </div>
-        ) : status ? (
-          <div className="space-y-5 animate-in fade-in-0 slide-in-from-bottom-1 duration-200">
-            {/* 并发状态概览 */}
-            <div className="grid grid-cols-3 gap-3">
-              <div className="rounded-xl bg-blue-50 dark:bg-blue-900/20 p-3 text-center">
-                <p className="text-2xl font-bold text-blue-600">{status.running.length}</p>
-                <p className="text-xs text-slate-500 mt-1">运行中</p>
-                <p className="text-xs text-blue-400 mt-0.5">上限 {status.concurrencyLimit}</p>
-              </div>
-              <div className="rounded-xl bg-amber-50 dark:bg-amber-900/20 p-3 text-center">
-                <p className="text-2xl font-bold text-amber-600">
-                  {(status.queueDepth ?? status.queued.length) + (status.directQueueDepth ?? status.directQueued.length)}
-                </p>
-                <p className="text-xs text-slate-500 mt-1">等待队列</p>
-                {status.maxQueueDepth && (
-                  <p className="text-xs text-amber-400 mt-0.5">上限 {status.maxQueueDepth}</p>
-                )}
-              </div>
-              <div className="rounded-xl bg-green-50 dark:bg-green-900/20 p-3 text-center">
-                <p className="text-2xl font-bold text-green-600">{status.scheduled.length}</p>
-                <p className="text-xs text-slate-500 mt-1">已计划</p>
-              </div>
-            </div>
-
-            {/* 并发使用率进度条 */}
-            <div className="space-y-1.5">
-              <div className="flex items-center justify-between text-xs text-slate-500">
-                <span>并发使用率</span>
-                <span>{status.running.length} / {status.concurrencyLimit}</span>
-              </div>
-              <Progress
-                value={Math.min(100, (status.running.length / Math.max(1, status.concurrencyLimit)) * 100)}
-                className={cn(
-                  'h-2 w-full',
-                  status.running.length >= status.concurrencyLimit
-                    ? '[&>div]:bg-red-500'
-                    : status.running.length >= status.concurrencyLimit * 0.7
-                    ? '[&>div]:bg-amber-400'
-                    : '[&>div]:bg-blue-500'
-                )}
-                aria-label={`并发使用率 ${status.running.length} / ${status.concurrencyLimit}`}
-              />
-            </div>
-
-            {/* 运行中的任务（P1：显示 taskId + runId + 已运行时长） */}
-            {status.running.length > 0 && (
-              <div className="space-y-1.5">
-                <h4 className="text-xs font-semibold text-slate-600 dark:text-slate-400 uppercase tracking-wider flex items-center gap-1.5">
-                  <span className="w-2 h-2 rounded-full bg-blue-500 animate-pulse inline-block" />
-                  运行中 ({status.running.length})
-                </h4>
-                <div className="flex flex-col gap-1.5">
-                  {status.running.map((slot) => {
-                    const elapsedSec = Math.round((slot.elapsedMs ?? 0) / 1000);
-                    const elapsedDisplay = elapsedSec >= 60
-                      ? `${Math.floor(elapsedSec / 60)}m${elapsedSec % 60}s`
-                      : `${elapsedSec}s`;
-                    return (
-                      <div key={slot.runId ?? slot.taskId} className="flex items-center justify-between px-2 py-1 rounded-md text-xs bg-blue-100 dark:bg-blue-900/30 text-blue-700 dark:text-blue-300 font-mono">
-                        <span>Task #{slot.taskId}</span>
-                        <span className="text-blue-400 ml-2">Run #{slot.runId}</span>
-                        <span className="text-blue-500 ml-auto">{elapsedDisplay}</span>
-                      </div>
-                    );
-                  })}
-                </div>
-              </div>
-            )}
-
-            {/* 等待队列（P1：显示优先级、等待时长、触发方式） */}
-            {status.queued.length > 0 && (
-              <div className="space-y-1.5">
-                <h4 className="text-xs font-semibold text-slate-600 dark:text-slate-400 uppercase tracking-wider flex items-center gap-1.5">
-                  <span className="w-2 h-2 rounded-full bg-amber-400 inline-block" />
-                  等待队列 ({status.queued.length})
-                </h4>
-                <div className="flex flex-col gap-1.5">
-                  {status.queued.map((item) => {
-                    const waitSec = Math.round((item.waitMs ?? 0) / 1000);
-                    const waitDisplay = waitSec >= 60
-                      ? `${Math.floor(waitSec / 60)}m${waitSec % 60}s`
-                      : `${waitSec}s`;
-                    const triggerLabel = item.triggerReason === 'manual' ? '手动' : item.triggerReason === 'retry' ? '重试' : '定时';
-                    return (
-                      <div key={`${item.taskId}-${item.queuePosition}`} className="flex items-center gap-2 px-2 py-1 rounded-md text-xs bg-amber-50 dark:bg-amber-900/20 text-amber-700 dark:text-amber-300 font-mono">
-                        <span className="text-amber-400 font-bold w-5">#{item.queuePosition}</span>
-                        <span>Task #{item.taskId}</span>
-                        <span className="ml-1 px-1 py-0.5 rounded bg-amber-100 dark:bg-amber-800/30 text-amber-600">{triggerLabel}</span>
-                        <span className="ml-auto text-amber-400">等待 {waitDisplay}</span>
-                      </div>
-                    );
-                  })}
-                </div>
-              </div>
-            )}
-
-            {/* 直连等待队列（run-case / run-batch） */}
-            {status.directQueued.length > 0 && (
-              <div className="space-y-1.5">
-                <h4 className="text-xs font-semibold text-slate-600 dark:text-slate-400 uppercase tracking-wider flex items-center gap-1.5">
-                  <span className="w-2 h-2 rounded-full bg-violet-400 inline-block" />
-                  直连等待队列 ({status.directQueued.length})
-                </h4>
-                <div className="flex flex-col gap-1.5">
-                  {status.directQueued.map((item) => {
-                    const waitSec = Math.round((item.waitMs ?? 0) / 1000);
-                    const waitDisplay = waitSec >= 60
-                      ? `${Math.floor(waitSec / 60)}m${waitSec % 60}s`
-                      : `${waitSec}s`;
-                    return (
-                      <div key={`${item.label}-${item.queuePosition}`} className="flex items-center gap-2 px-2 py-1 rounded-md text-xs bg-violet-50 dark:bg-violet-900/20 text-violet-700 dark:text-violet-300 font-mono">
-                        <span className="text-violet-400 font-bold w-5">#{item.queuePosition}</span>
-                        <span className="truncate">{item.label}</span>
-                        <span className="ml-auto text-violet-400">等待 {waitDisplay}</span>
-                      </div>
-                    );
-                  })}
-                </div>
-              </div>
-            )}
-
-            {/* 已计划的任务 */}
-            {status.scheduled.length > 0 && (
-              <div className="space-y-1.5">
-                <h4 className="text-xs font-semibold text-slate-600 dark:text-slate-400 uppercase tracking-wider flex items-center gap-1.5">
-                  <span className="w-2 h-2 rounded-full bg-green-500 inline-block" />
-                  已计划定时触发 ({status.scheduled.length})
-                </h4>
-                <div className="flex flex-wrap gap-1.5">
-                  {status.scheduled.map((id) => (
-                    <span key={id} className="px-2 py-1 rounded-md text-xs bg-green-50 dark:bg-green-900/20 text-green-700 dark:text-green-300 font-mono">
-                      Task #{id}
-                    </span>
-                  ))}
-                </div>
-              </div>
-            )}
-
-            {status.running.length === 0 && status.queued.length === 0 && status.directQueued.length === 0 && (
-              <div className="text-center py-4 text-slate-400">
-                <CheckCircle2 className="h-10 w-10 mx-auto mb-2 opacity-30" />
-                <p className="text-sm">调度器空闲，无任务运行</p>
-              </div>
-            )}
-          </div>
-        ) : null}
-        </div>
-
-        <DialogFooter className="gap-2 px-6 py-3 border-t border-slate-100 dark:border-slate-800 bg-slate-50/80 dark:bg-slate-900/60">
-          <Button variant="ghost" size="sm" onClick={() => refetch()} className="gap-1.5">
-            <RefreshCw className="h-3.5 w-3.5" />
-            刷新
-          </Button>
-          <Button variant="outline" onClick={onClose}>关闭</Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-/* ─── 新建 / 编辑 弹窗 ───────────────────────────────────── */
-
-interface TaskFormDialogProps {
-  open: boolean;
-  task: Task | null;
-  onClose: () => void;
-  onSave: (input: CreateTaskInput & { id?: number }) => void;
-  isSaving: boolean;
-}
-
-/** 用例类型标签映射 */
-const CASE_TYPE_LABELS: Record<string, string> = {
-  api: 'API',
-  ui: 'UI',
-  performance: '性能',
-};
-
-/** 用例类型颜色 */
-const CASE_TYPE_COLORS: Record<string, string> = {
-  api: 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300',
-  ui: 'bg-purple-100 text-purple-700 dark:bg-purple-900/30 dark:text-purple-300',
-  performance: 'bg-orange-100 text-orange-700 dark:bg-orange-900/30 dark:text-orange-300',
-};
-
-/** Cron 常用预设 */
-const CRON_PRESETS = [
-  { label: '每分钟', value: '* * * * *' },
-  { label: '每小时', value: '0 * * * *' },
-  { label: '每天凌晨', value: '0 2 * * *' },
-  { label: '工作日9点', value: '0 9 * * 1-5' },
-  { label: '每周一', value: '0 9 * * 1' },
-  { label: '每月1号', value: '0 9 1 * *' },
-] as const;
-
-function TaskFormDialog({ open, task, onClose, onSave, isSaving }: TaskFormDialogProps) {
-  const [name, setName] = useState('');
-  const [description, setDescription] = useState('');
-  const [triggerType, setTriggerType] = useState<TaskTriggerType>('manual');
-  const [cronExpression, setCronExpression] = useState('');
-  const [errors, setErrors] = useState<Record<string, string>>({});
-
-  // ── Cron 预览（防抖后才触发查询）──────────────────────────────────
-  const [debouncedCron, setDebouncedCron] = useState('');
-  useEffect(() => {
-    const t = setTimeout(() => setDebouncedCron(cronExpression), 600);
-    return () => clearTimeout(t);
-  }, [cronExpression]);
-
-  const cronPreviewExpr = triggerType === 'scheduled' ? debouncedCron : '';
-  const {
-    data: cronPreviewData,
-    isFetching: cronPreviewLoading,
-  } = useCronPreview(cronPreviewExpr, 5);
-
-  // ── 关联用例状态 ──────────────────────────────────────
-  const [selectedCaseIds, setSelectedCaseIds] = useState<number[]>([]);
-  const [caseSearch, setCaseSearch] = useState('');
-  const [caseTypeFilter, setCaseTypeFilter] = useState<CaseType | ''>('');
-  const [casePickerOpen, setCasePickerOpen] = useState(false);
-
-  // 已选用例对象缓存（独立维护，避免 API 加载中时标签短暂消失）
-  const [selectedCaseObjects, setSelectedCaseObjects] = useState<CaseItem[]>([]);
-
-  // 搜索防抖
-  const [debouncedSearch, setDebouncedSearch] = useState('');
-  useEffect(() => {
-    const t = setTimeout(() => setDebouncedSearch(caseSearch), 300);
-    return () => clearTimeout(t);
-  }, [caseSearch]);
-
-  const { data: casesData, isLoading: casesLoading } = useAllCasesForSelect({
-    search: debouncedSearch,
-    type: caseTypeFilter,
-    enabled: open, // 弹窗打开时才请求
-  });
-  const allCases = casesData?.data ?? [];
-
-  // 同步已选用例对象缓存（避免在事件回调内嵌套 setState）
-  useEffect(() => {
-    if (selectedCaseIds.length === 0) {
-      setSelectedCaseObjects([]);
-      return;
-    }
-
-    setSelectedCaseObjects((prev) => {
-      const cacheById = new Map<number, CaseItem>();
-      prev.forEach((item) => cacheById.set(item.id, item));
-      allCases.forEach((item) => cacheById.set(item.id, item));
-
-      return selectedCaseIds
-        .map((id) => cacheById.get(id))
-        .filter((item): item is CaseItem => item !== undefined);
-    });
-  }, [allCases, selectedCaseIds]);
-
-  // 候选列表（排除已选）
-  const candidateCases = useMemo(
-    () => allCases.filter((c) => !selectedCaseIds.includes(c.id)),
-    [allCases, selectedCaseIds]
-  );
-
-  // 已选用例对象：按 selectedCaseIds 顺序展示，保证顺序稳定
-  const selectedCases = selectedCaseObjects;
-
-  // 回填编辑数据
-  useEffect(() => {
-    if (open) {
-      setName(task?.name ?? '');
-      setDescription(task?.description ?? '');
-      setTriggerType(task?.trigger_type ?? 'manual');
-      setCronExpression(task?.cron_expression ?? '');
-      setErrors({});
-      setCaseSearch('');
-      setCaseTypeFilter('');
-      setCasePickerOpen(false);
-      setSelectedCaseObjects([]);
-      // 回填已有 case_ids
-      if (task?.case_ids) {
-        try {
-          const ids = JSON.parse(task.case_ids);
-          setSelectedCaseIds(Array.isArray(ids) ? ids : []);
-        } catch {
-          setSelectedCaseIds([]);
-        }
-      } else {
-        setSelectedCaseIds([]);
-      }
-    }
-  }, [open, task]);
-
-  const toggleCaseSelect = useCallback((caseId: number) => {
-    setSelectedCaseIds((prev) =>
-      prev.includes(caseId)
-        ? prev.filter((id) => id !== caseId)
-        : [...prev, caseId]
-    );
-  }, []);
-
-  const validate = () => {
-    const errs: Record<string, string> = {};
-
-    // 任务名称验证
-    if (!name.trim()) {
-      errs.name = TASK_MESSAGES.NAME_REQUIRED;
-    } else if (name.trim().length > TASKS_CONFIG.MAX_NAME_LENGTH) {
-      errs.name = TASK_MESSAGES.NAME_TOO_LONG;
-    } else if (!/^[\u4e00-\u9fa5a-zA-Z0-9_\-\s]+$/.test(name.trim())) {
-      errs.name = TASK_MESSAGES.NAME_INVALID_CHARS;
-    }
-
-    // 描述验证
-    if (description.trim().length > TASKS_CONFIG.MAX_DESCRIPTION_LENGTH) {
-      errs.description = TASK_MESSAGES.DESCRIPTION_TOO_LONG;
-    }
-
-    // Cron 表达式验证
-    if (triggerType === 'scheduled') {
-      if (!cronExpression.trim()) {
-        errs.cronExpression = TASK_MESSAGES.CRON_REQUIRED;
-      } else {
-        const parts = cronExpression.trim().split(/\s+/);
-        if (parts.length !== TASKS_CONFIG.CRON_SEGMENTS) {
-          errs.cronExpression = TASK_MESSAGES.CRON_INVALID_FORMAT;
-        } else {
-          // 验证每段的合法性（简单验证：数字、*、逗号、连字符、斜杠）
-          const isValid = parts.every((part) => /^[\d\*\-,\/]+$/.test(part));
-          if (!isValid) {
-            errs.cronExpression = TASK_MESSAGES.CRON_INVALID_CHARS;
-          }
-        }
-      }
-    }
-
-    return errs;
-  };
-
-  const handleSubmit = () => {
-    const errs = validate();
-    if (Object.keys(errs).length > 0) {
-      setErrors(errs);
-      return;
-    }
-    onSave({
-      id: task?.id,
-      name: name.trim(),
-      description: description.trim() || undefined,
-      triggerType,
-      cronExpression: triggerType === 'scheduled' ? cronExpression.trim() : undefined,
-      caseIds: selectedCaseIds,
-    });
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={(v) => !v && onClose()}>
-      <DialogContent className="max-w-3xl max-h-[90vh] overflow-y-auto p-0 animate-in fade-in-0 zoom-in-95 duration-200">
-        <DialogHeader className="px-6 pt-6 pb-3 border-b border-slate-100 dark:border-slate-800">
-          <DialogTitle>{task ? TASK_MESSAGES.FORM_EDIT_TITLE : TASK_MESSAGES.FORM_CREATE_TITLE}</DialogTitle>
-          <DialogDescription>
-            {task ? TASK_MESSAGES.FORM_EDIT_DESC : TASK_MESSAGES.FORM_CREATE_DESC}
-          </DialogDescription>
-        </DialogHeader>
-
-        <div className="space-y-5 px-6 py-4">
-          {/* 任务名称 */}
-          <div className="space-y-1.5">
-            <label className="text-sm font-medium text-slate-700 dark:text-slate-300">
-              {TASK_MESSAGES.FORM_NAME_LABEL} <span className="text-red-500">*</span>
-            </label>
-            <Input
-              placeholder={TASK_MESSAGES.FORM_NAME_PLACEHOLDER}
-              value={name}
-              onChange={(e) => {
-                setName(e.target.value);
-                if (errors.name) setErrors((p) => ({ ...p, name: '' }));
-              }}
-              className={cn(errors.name && 'border-red-400 focus-visible:ring-red-400')}
-              aria-label={TASK_MESSAGES.FORM_NAME_LABEL}
-              aria-invalid={!!errors.name}
-              aria-describedby={errors.name ? 'name-error' : undefined}
-            />
-            {errors.name && <p id="name-error" className="text-xs text-red-500">{errors.name}</p>}
-          </div>
-
-          {/* 任务描述 */}
-          <div className="space-y-1.5">
-            <label className="text-sm font-medium text-slate-700 dark:text-slate-300">
-              {TASK_MESSAGES.FORM_DESCRIPTION_LABEL}
-            </label>
-            <Textarea
-              placeholder={TASK_MESSAGES.FORM_DESCRIPTION_PLACEHOLDER}
-              value={description}
-              onChange={(e) => {
-                setDescription(e.target.value);
-                if (errors.description) setErrors((p) => ({ ...p, description: '' }));
-              }}
-              rows={3}
-              className={cn(errors.description && 'border-red-400 focus-visible:ring-red-400')}
-              aria-label={TASK_MESSAGES.FORM_DESCRIPTION_LABEL}
-              aria-invalid={!!errors.description}
-              aria-describedby={errors.description ? 'description-error' : undefined}
-            />
-            {errors.description && <p id="description-error" className="text-xs text-red-500">{errors.description}</p>}
-          </div>
-
-          {/* 触发类型 */}
-          <div className="space-y-1.5">
-            <label className="text-sm font-medium text-slate-700 dark:text-slate-300">
-              {TASK_MESSAGES.FORM_TRIGGER_LABEL}
-            </label>
-            <div className="flex gap-2" role="group" aria-label={TASK_MESSAGES.FORM_TRIGGER_LABEL}>
-              {TASK_TRIGGER_TYPE_OPTIONS.map(({ value, label }) => (
-                <button
-                  key={value}
-                  type="button"
-                  onClick={() => setTriggerType(value)}
-                  className={cn(
-                    'flex-1 rounded-lg border px-3 py-2.5 text-sm font-medium transition-all',
-                    triggerType === value
-                      ? 'border-blue-500 bg-blue-50 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300 dark:border-blue-500'
-                      : 'border-slate-200 text-slate-600 hover:border-slate-300 dark:border-slate-700 dark:text-slate-400'
-                  )}
-                  aria-pressed={triggerType === value}
-                >
-                  {label}
-                </button>
-              ))}
-            </div>
-          </div>
-
-          {/* Cron 表达式（仅定时触发时显示） */}
-          {triggerType === 'scheduled' && (
-            <div className="space-y-2">
-              <label className="text-sm font-medium text-slate-700 dark:text-slate-300">
-                {TASK_MESSAGES.FORM_CRON_LABEL} <span className="text-red-500">*</span>
-              </label>
-
-              {/* 常用预设快捷按钮 */}
-              <div className="flex flex-wrap gap-1.5">
-                {CRON_PRESETS.map((preset) => (
-                  <button
-                    key={preset.value}
-                    type="button"
-                    onClick={() => {
-                      setCronExpression(preset.value);
-                      if (errors.cronExpression) setErrors((p) => ({ ...p, cronExpression: '' }));
-                    }}
-                    className={cn(
-                      'px-2.5 py-1 rounded-full text-xs font-medium border transition-all',
-                      cronExpression === preset.value
-                        ? 'border-blue-500 bg-blue-50 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300 dark:border-blue-500'
-                        : 'border-slate-200 text-slate-500 hover:border-blue-300 hover:text-blue-600 dark:border-slate-700 dark:text-slate-400'
-                    )}
-                  >
-                    {preset.label}
-                  </button>
-                ))}
-              </div>
-
-              {/* Cron 输入框 */}
-              <Input
-                placeholder={TASK_MESSAGES.FORM_CRON_PLACEHOLDER}
-                value={cronExpression}
-                onChange={(e) => {
-                  setCronExpression(e.target.value);
-                  if (errors.cronExpression) setErrors((p) => ({ ...p, cronExpression: '' }));
-                }}
-                className={cn(
-                  'font-mono',
-                  errors.cronExpression && 'border-red-400 focus-visible:ring-red-400'
-                )}
-                aria-label={TASK_MESSAGES.FORM_CRON_LABEL}
-                aria-invalid={!!errors.cronExpression}
-                aria-describedby={errors.cronExpression ? 'cron-error' : 'cron-hint'}
-              />
-              {errors.cronExpression ? (
-                <p id="cron-error" className="text-xs text-red-500">{errors.cronExpression}</p>
-              ) : (
-                <p id="cron-hint" className="text-xs text-slate-400">{TASK_MESSAGES.FORM_CRON_HINT}</p>
-              )}
-
-              {/* 下次运行时间预览卡片 */}
-              {(cronPreviewLoading || (cronPreviewData?.times && cronPreviewData.times.length > 0)) && (
-                <div className="rounded-lg border border-blue-100 dark:border-blue-900/40 bg-blue-50/60 dark:bg-blue-900/10 p-3 space-y-1.5">
-                  <div className="flex items-center gap-1.5 text-xs font-semibold text-blue-700 dark:text-blue-300">
-                    <Calendar className="h-3.5 w-3.5" />
-                    <span>未来运行时间预览</span>
-                    {cronPreviewLoading && <Loader2 className="h-3 w-3 animate-spin ml-auto" />}
-                  </div>
-                  {!cronPreviewLoading && cronPreviewData?.times.map((isoTime, idx) => {
-                    const d = new Date(isoTime);
-                    const dateStr = d.toLocaleDateString('zh-CN', {
-                      month: '2-digit', day: '2-digit', weekday: 'short',
-                    });
-                    const timeStr = d.toLocaleTimeString('zh-CN', {
-                      hour: '2-digit', minute: '2-digit',
-                    });
-                    return (
-                      <div
-                        key={isoTime}
-                        className="flex items-center gap-2 text-xs text-slate-600 dark:text-slate-400"
-                      >
-                        <span className="inline-flex items-center justify-center w-4 h-4 rounded-full bg-blue-100 dark:bg-blue-900/40 text-blue-600 dark:text-blue-300 text-[10px] font-bold shrink-0">
-                          {idx + 1}
-                        </span>
-                        <span className="font-mono tabular-nums">{dateStr} {timeStr}</span>
-                      </div>
-                    );
-                  })}
-                </div>
-              )}
-            </div>
-          )}
-
-          {/* ── 关联用例 ─────────────────────────────────── */}
-          <div className="space-y-1.5">
-            <div className="flex items-center justify-between">
-              <label className="text-sm font-medium text-slate-700 dark:text-slate-300 flex items-center gap-1.5">
-                <ListChecks className="h-4 w-4 text-slate-500" />
-                关联用例
-                {selectedCaseIds.length > 0 && (
-                  <span className="ml-1 inline-flex items-center rounded-full bg-blue-100 px-2 py-0.5 text-xs font-semibold text-blue-700 dark:bg-blue-900/40 dark:text-blue-300">
-                    {selectedCaseIds.length}
-                  </span>
-                )}
-              </label>
-              <button
-                type="button"
-                onClick={() => setCasePickerOpen((v) => !v)}
-                className="text-xs text-blue-600 hover:text-blue-700 flex items-center gap-0.5 transition-colors"
-              >
-                {casePickerOpen ? '收起' : '展开选择'}
-                <ChevronDown className={cn('h-3.5 w-3.5 transition-transform', casePickerOpen && 'rotate-180')} />
-              </button>
-            </div>
-
-            {/* 已选用例标签列表 */}
-            {selectedCases.length > 0 && (
-              <div className="flex flex-wrap gap-1.5 p-2 rounded-lg border border-slate-200 dark:border-slate-700 bg-slate-50 dark:bg-slate-900/50 min-h-[36px]">
-                {selectedCases.map((c) => (
-                  <span
-                    key={c.id}
-                    className="inline-flex items-center gap-1 rounded-md bg-white dark:bg-slate-800 border border-slate-200 dark:border-slate-700 px-2 py-0.5 text-xs text-slate-700 dark:text-slate-300 shadow-sm"
-                  >
-                    <span className={cn('rounded px-1 text-[10px] font-semibold', CASE_TYPE_COLORS[c.type] ?? '')}>
-                      {CASE_TYPE_LABELS[c.type] ?? c.type}
-                    </span>
-                    <span className="max-w-[120px] truncate">{c.name}</span>
-                    <button
-                      type="button"
-                      onClick={() => toggleCaseSelect(c.id)}
-                      className="ml-0.5 text-slate-400 hover:text-red-500 transition-colors"
-                      aria-label={`移除 ${c.name}`}
-                    >
-                      <X className="h-3 w-3" />
-                    </button>
-                  </span>
-                ))}
-              </div>
-            )}
-
-            {selectedCases.length === 0 && !casePickerOpen && (
-              <p className="text-xs text-slate-400 italic">暂未关联用例，任务运行时将跳过执行</p>
-            )}
-
-            {/* 用例选择器面板 */}
-            {casePickerOpen && (
-              <div className="rounded-lg border border-slate-200 dark:border-slate-700 bg-white dark:bg-slate-900 shadow-sm animate-in fade-in-0 slide-in-from-top-1 duration-200">
-                {/* 搜索 + 类型过滤 */}
-                <div className="flex gap-2 p-2 border-b border-slate-100 dark:border-slate-800">
-                  <div className="relative flex-1">
-                    <Search className="absolute left-2.5 top-1/2 -translate-y-1/2 h-3.5 w-3.5 text-slate-400" />
-                    <input
-                      type="text"
-                      placeholder="搜索用例名称..."
-                      value={caseSearch}
-                      onChange={(e) => setCaseSearch(e.target.value)}
-                      className="w-full pl-8 pr-3 py-1.5 text-sm rounded-md border border-slate-200 dark:border-slate-700 bg-slate-50 dark:bg-slate-800 text-slate-700 dark:text-slate-300 placeholder-slate-400 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                    />
-                  </div>
-                  <div className="flex gap-1">
-                    {CASE_TYPE_FILTER_OPTIONS.map(({ value, label }) => (
-                      <button
-                        key={value}
-                        type="button"
-                        onClick={() => setCaseTypeFilter(value)}
-                        className={cn(
-                          'px-2 py-1 rounded-md text-xs font-medium border transition-all',
-                          caseTypeFilter === value
-                            ? 'border-blue-500 bg-blue-50 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300'
-                            : 'border-slate-200 text-slate-500 hover:border-slate-300 dark:border-slate-700 dark:text-slate-400'
-                        )}
-                      >
-                        {label}
-                      </button>
-                    ))}
-                  </div>
-                </div>
-
-                {/* 候选用例列表 */}
-                <div className="max-h-[200px] overflow-y-auto">
-                  {casesLoading ? (
-                    <div className="flex items-center justify-center py-6 text-slate-400">
-                      <Loader2 className="h-4 w-4 animate-spin mr-2" />
-                      <span className="text-sm">加载中...</span>
-                    </div>
-                  ) : candidateCases.length === 0 ? (
-                    <div className="flex flex-col items-center justify-center py-6 text-slate-400">
-                      <Search className="h-5 w-5 mb-1 opacity-40" />
-                      <span className="text-sm">{caseSearch ? '未找到匹配用例' : '所有用例已选完'}</span>
-                    </div>
-                  ) : (
-                    candidateCases.map((c) => (
-                      <button
-                        key={c.id}
-                        type="button"
-                        onClick={() => toggleCaseSelect(c.id)}
-                        className="w-full flex items-center gap-2 px-3 py-2 text-left hover:bg-slate-50 dark:hover:bg-slate-800 transition-colors border-b border-slate-50 dark:border-slate-800 last:border-0"
-                      >
-                        <span className={cn('shrink-0 rounded px-1.5 py-0.5 text-[10px] font-bold', CASE_TYPE_COLORS[c.type] ?? '')}>
-                          {CASE_TYPE_LABELS[c.type] ?? c.type}
-                        </span>
-                        <span className="flex-1 text-sm text-slate-700 dark:text-slate-300 truncate">{c.name}</span>
-                        {c.module && (
-                          <span className="shrink-0 text-[10px] text-slate-400 bg-slate-100 dark:bg-slate-700 rounded px-1.5 py-0.5 truncate max-w-[80px]">
-                            {c.module}
-                          </span>
-                        )}
-                        <Plus className="shrink-0 h-3.5 w-3.5 text-blue-500" />
-                      </button>
-                    ))
-                  )}
-                </div>
-
-                {/* 底部统计 */}
-                <div className="flex items-center justify-between px-3 py-1.5 border-t border-slate-100 dark:border-slate-800 bg-slate-50/60 dark:bg-slate-900/60">
-                  <span className="text-[11px] text-slate-400">
-                    共 {casesData?.total ?? 0} 个用例，已选 {selectedCaseIds.length} 个
-                  </span>
-                  {selectedCaseIds.length > 0 && (
-                    <button
-                      type="button"
-                      onClick={() => {
-                        setSelectedCaseIds([]);
-                        setSelectedCaseObjects([]);
-                      }}
-                      className="text-[11px] text-red-400 hover:text-red-600 transition-colors"
-                    >
-                      清空已选
-                    </button>
-                  )}
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-
-        <DialogFooter className="px-6 py-4 border-t border-slate-100 dark:border-slate-800 bg-slate-50/80 dark:bg-slate-900/60">
-          <Button variant="outline" onClick={onClose} disabled={isSaving}>
-            {TASK_MESSAGES.BTN_CANCEL}
-          </Button>
-          <Button onClick={handleSubmit} disabled={isSaving}>
-            {isSaving ? <Loader2 className="h-4 w-4 animate-spin mr-2" /> : null}
-            {isSaving ? TASK_MESSAGES.BTN_SAVING : task ? TASK_MESSAGES.BTN_SAVE : TASK_MESSAGES.BTN_CREATE_TASK}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
   );
 }
